# docs/openapi_client.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Caser Client API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/cases/{case_id}/narratives": {
      "post": {
        "tags": [
          "narratives"
        ],
        "summary": "Generate Narrative",
        "description": "Fixed-template generation (CLAUDE.md: no real AI/LLM needed). Always\ninserts a new row — never edits an existing one; the newest row for a\ncase is the current/authoritative one, older rows stay as history (same\n\"rows accumulate, newest wins\" pattern as Subscriptions). period/language\nare chosen by the lawyer at generation time and stored as a fixed\nsnapshot on the row, same reasoning as total_hours/total_fee.",
        "operationId": "generate_narrative_cases__case_id__narratives_post",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateNarrativeRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NarrativeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "narratives"
        ],
        "summary": "List Narratives",
        "description": "Full history, newest first — the first row returned is always the\ncurrent/authoritative narrative for the case.",
        "operationId": "list_narratives_cases__case_id__narratives_get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Case Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_NarrativeResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cases/{case_id}/narratives/{narrative_id}/export-pdf": {
      "post": {
        "tags": [
          "narratives"
        ],
        "summary": "Export Narrative Pdf",
        "description": "Renders the narrative to PDF and files it as a real Document on the\ncase, internal folder by default (CLAUDE.md). This is what makes a\nnarrative reachable by the client at all — sharing it afterwards is just\nthe existing Documents reclassify-to-client-folder action, no separate\nvisibility mechanism needed here.",
        "operationId": "export_narrative_pdf_cases__case_id__narratives__narrative_id__export_pdf_post",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Case Id"
            }
          },
          {
            "name": "narrative_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Narrative Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cases/{case_id}/work-logs": {
      "post": {
        "tags": [
          "work-logs"
        ],
        "summary": "Create Work Log",
        "description": "Manual entry — the case must be one this lawyer is currently assigned\nto, and not closed (CLAUDE.md: a closed case blocks new WorkLogs the same\nway it blocks new Documents).",
        "operationId": "create_work_log_cases__case_id__work_logs_post",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkLogRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkLogResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "work-logs"
        ],
        "summary": "List Work Logs",
        "operationId": "list_work_logs_cases__case_id__work_logs_get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Case Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_WorkLogResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cases/{case_id}/work-logs/{work_log_id}": {
      "patch": {
        "tags": [
          "work-logs"
        ],
        "summary": "Update Work Log",
        "description": "Any lawyer assigned to the case can edit any entry on it, not just\ntheir own (CLAUDE.md: same broad, collaborative authority as Documents).\nLocked entirely once the case is closed — same cutoff as creating new\nentries, one consistent \"closed means frozen\" rule. Every edit is logged\nin AuditLogs (who changed which entry, when — not before/after values).",
        "operationId": "update_work_log_cases__case_id__work_logs__work_log_id__patch",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Case Id"
            }
          },
          {
            "name": "work_log_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Work Log Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkLogRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkLogResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "work-logs"
        ],
        "summary": "Delete Work Log",
        "operationId": "delete_work_log_cases__case_id__work_logs__work_log_id__delete",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Case Id"
            }
          },
          {
            "name": "work_log_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Work Log Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/work-logs/import": {
      "post": {
        "tags": [
          "work-logs"
        ],
        "summary": "Import Work Logs",
        "description": "Self-import — every row is this lawyer's own hours, no lawyer-email\ncolumn (see admin_api's variant for the bulk, on-someone-else's-behalf\nversion). Whole-file validation: any row failing rejects the entire file,\nnothing is written (shared/worklog_import.py enforces the actual rules).",
        "operationId": "import_work_logs_work_logs_import_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_work_logs_work_logs_import_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkLogImportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_import_work_logs_work_logs_import_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_import_work_logs_work_logs_import_post"
      },
...
      "CreateWorkLogRequest": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "hours": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 24.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Hours"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "date",
          "hours"
        ],
        "title": "CreateWorkLogRequest"
      },
      "DocumentFolderType": {
        "type": "string",
        "enum": [
          "client",
          "internal"
        ],
        "title": "DocumentFolderType"
      },
      "DocumentResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "case_id": {
            "type": "integer",
            "title": "Case Id"
          },
          "folder_type": {
            "$ref": "#/components/schemas/DocumentFolderType"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "file_size": {
            "type": "integer",
            "title": "File Size"
          },
          "uploaded_by": {
            "type": "integer",
            "title": "Uploaded By"
          },
          "uploader_name": {
            "type": "string",
            "title": "Uploader Name"
          },
          "archived_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "case_id",
          "folder_type",
          "original_filename",
          "content_type",
          "file_size",
          "uploaded_by",
          "uploader_name",
          "created_at"
        ],
        "title": "DocumentResponse"
      },
...
      "GenerateNarrativeRequest": {
        "properties": {
          "period_start": {
            "type": "string",
            "format": "date",
            "title": "Period Start"
          },
          "period_end": {
            "type": "string",
            "format": "date",
            "title": "Period End"
          },
          "language": {
            "$ref": "#/components/schemas/NarrativeLanguage",
            "default": "he"
          }
        },
        "type": "object",
        "required": [
          "period_start",
          "period_end"
        ],
        "title": "GenerateNarrativeRequest",
        "description": "period_start/period_end are chosen by the lawyer at generation time —\nreal legal billing is period-by-period, not every hour ever logged on\nthe case (CLAUDE.md). No overlap-prevention against other narratives on\nthe same case: re-covering a period already billed is a legitimate\ncorrection, not an error."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NarrativeLanguage": {
        "type": "string",
        "enum": [
          "he",
          "en"
        ],
        "title": "NarrativeLanguage"
      },
      "NarrativeResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "case_id": {
            "type": "integer",
            "title": "Case Id"
          },
          "generated_text": {
            "type": "string",
            "title": "Generated Text"
          },
          "total_hours": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Hours"
          },
          "total_fee": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Fee"
          },
          "language": {
            "$ref": "#/components/schemas/NarrativeLanguage"
          },
          "period_start": {
            "type": "string",
            "format": "date",
            "title": "Period Start"
          },
          "period_end": {
            "type": "string",
            "format": "date",
            "title": "Period End"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "case_id",
          "generated_text",
          "total_hours",
          "total_fee",
          "language",
          "period_start",
          "period_end",
          "created_at"
        ],
        "title": "NarrativeResponse"
      },
...
      "Page_NarrativeResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NarrativeResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[NarrativeResponse]"
      },
      "Page_WorkLogResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WorkLogResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[WorkLogResponse]"
      },
...
      "UpdateWorkLogRequest": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "hours": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 24.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Hours"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "date",
          "hours"
        ],
        "title": "UpdateWorkLogRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkLogImportResponse": {
        "properties": {
          "imported_count": {
            "type": "integer",
            "title": "Imported Count"
          }
        },
        "type": "object",
        "required": [
          "imported_count"
        ],
        "title": "WorkLogImportResponse"
      },
      "WorkLogResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "case_id": {
            "type": "integer",
            "title": "Case Id"
          },
          "lawyer_id": {
            "type": "integer",
            "title": "Lawyer Id"
          },
          "lawyer_name": {
            "type": "string",
            "title": "Lawyer Name"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "hours": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Hours"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "source": {
            "$ref": "#/components/schemas/WorkLogSource"
          }
        },
        "type": "object",
        "required": [
          "id",
          "case_id",
          "lawyer_id",
          "lawyer_name",
          "date",
          "hours",
          "source"
        ],
        "title": "WorkLogResponse"
      },
      "WorkLogSource": {
        "type": "string",
        "enum": [
          "manual",
          "excel_import"
        ],
        "title": "WorkLogSource"
      }
    }
  }
}